Fix hash(Clifford). Remove check in Pauli constructor.
hash(Clifford) was doing a reduction of the key and value lists, which reduced
them to single Paulis. This was bad...

Dropping the check in the Pauli constructor was because it appeared to be the
performance bottleneck in 2Q RB tests. Ideally this would become unnecessary
after creating the appropriate 2-bit bitstype for the Pauli vector.